@@ -1,25 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <UserSecretsId>aspnet-ManagamentApp-3bcf4517-84af-405c-906c-6c49179612dc</UserSecretsId>
     <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <DockerfileContext>..\..</DockerfileContext>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="8.0.6" />
-    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.6" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.20.1" />
-  </ItemGroup>
-
   <ItemGroup>
     <ProjectReference Include="..\Common\Application\BIManagement.Common.Application.csproj" />
     <ProjectReference Include="..\Common\Components\BIManagement.Common.Components.csproj" />
@@ -27,15 +14,21 @@
     <ProjectReference Include="..\Common\Persistence\BIManagement.Common.Persistence.csproj" />
     <ProjectReference Include="..\Common\Shared\BIManagement.Common.Shared.csproj" />
     <ProjectReference Include="..\Modules\DataIntegration\Infrastructure\BIManagement.Modules.DataIntegration.Infrastructure.csproj" />
-    <ProjectReference Include="..\Modules\DataIntegration\MapperPages\BIManagement.Modules.DataIntegration.MapperComponent.csproj" />
     <ProjectReference Include="..\Modules\DataIntegration\Pages\BIManagement.Modules.DataIntegration.Pages.csproj" />
     <ProjectReference Include="..\Modules\Deployment\Infrastructure\BIManagement.Modules.Deployment.Infrastructure.csproj" />
     <ProjectReference Include="..\Modules\Deployment\Pages\BIManagement.Modules.Deployment.Pages.csproj" />
     <ProjectReference Include="..\Modules\Notifications\Infrastructure\BIManagement.Modules.Notifications.Infrastructure.csproj" />
     <ProjectReference Include="..\Modules\Notifications\Pages\BIManagement.Modules.Notifications.Pages.csproj" />
+    <ProjectReference Include="..\Modules\Users\Api\BIManagement.Modules.Users.Api.csproj" />
     <ProjectReference Include="..\Modules\Users\Infrastructure\BIManagement.Modules.Users.Infrastructure.csproj" />
     <ProjectReference Include="..\Modules\Users\Pages\BIManagement.Modules.Users.Pages.csproj" />
     <ProjectReference Include="..\Modules\Users\Persistence\BIManagement.Modules.Users.Persistence.csproj" />
   </ItemGroup>
-
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="9.0.0-preview.5.24306.11" />
+    <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="9.0.0-preview.5.24306.11" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="9.0.0-preview.5.24306.3" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.0-preview.5.24306.3" />
+    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.21.0" />
+  </ItemGroup>
 </Project>
