@@ -1,4 +1,5 @@
 ﻿using BIManagement.Modules.DataIntegration.Application.Mapping.SqlViewGenerating;
+using BIManagement.Modules.DataIntegration.Domain.DbModelling;
 using BIManagement.Modules.DataIntegration.Domain.Mapping.JsonModel;
 using BIManagement.Modules.DataIntegration.Domain.Mapping.JsonModel.SourceEntities;
 using BIManagement.Modules.DataIntegration.Domain.Mapping.JsonModel.SourceEntities.Agregators;
@@ -9,69 +10,113 @@ namespace BIManagement.Test.Modules.DataIntegration.SqlViewGeneratorTests;
 [TestFixture]
 public class SqlViewVisitorTests
 {
+    [Test]
+    public void TestJoinProcessing()    // TODO: Test different join types and join condition operators
+    {
+        SourceColumn[] sourceColumns = [
+                   new("MzdObd_DatumOd", new SimpleType(SimpleType.Types.Date, false)),
+            new("MzdObd_DatumDo", new SimpleType(SimpleType.Types.Date, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ];
+
+        ISourceEntity sourceTable1 = new SourceTable("TabMzdList", null, [
+            new("ZamestnanecId", new SimpleType(SimpleType.Types.Integer, false)),
+            new("OdpracHod", new SimpleType(SimpleType.Types.Decimal, false)),
+            new("HodSaz", new SimpleType(SimpleType.Types.Decimal, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ]);
+
+        ISourceEntity sourceTable2 = new SourceTable("TabMzdObd", null, [
+            new("MzdObd_DatumOd", new SimpleType(SimpleType.Types.Date, false)),
+            new("MzdObd_DatumDo", new SimpleType(SimpleType.Types.Date, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ]);
+        Join join = new(
+            Join.Type.inner,
+            sourceTable1,
+            sourceTable2,
+            "join1",
+            [
+                sourceTable1.GetColumnMapping("IdObdobi"),
+                sourceTable1.GetColumnMapping("HodSaz"),
+                sourceTable1.GetColumnMapping("OdpracHod"),
+                sourceTable1.GetColumnMapping("ZamestnanecId"),
+                sourceTable2.GetColumnMapping("IdObdobi"),
+                sourceTable2.GetColumnMapping("MzdObd_DatumDo"),
+                sourceTable2.GetColumnMapping("MzdObd_DatumOd"),
+            ],
+            new(JoinCondition.Operator.equals, sourceTable2.GetColumnMapping("IdObdobi"), sourceTable2.GetColumnMapping("IdObdobi")));
+        join.AssignColumnOwnership();
+
+        SqlViewVisitor visitor = new("[dbo].");
+        join.Accept(visitor);
+        string actualJoinRepresentation = visitor.GetSqlView();
+
+        Assert.That(actualJoinRepresentation, Is.Not.Empty);
+        Assert.That(actualJoinRepresentation, Is.EqualTo("(SELECT TabMzdList.TabMzdList__IdObdobi,\r\nTabMzdList.TabMzdList__HodSaz,\r\nTabMzdList.TabMzdList__OdpracHod,\r\nTabMzdList.TabMzdList__ZamestnanecId,\r\nTabMzdObd.TabMzdObd__IdObdobi,\r\nTabMzdObd.TabMzdObd__MzdObd_DatumDo,\r\nTabMzdObd.TabMzdObd__MzdObd_DatumOd\r\nFROM (SELECT TabMzdList.ZamestnanecId AS TabMzdList__ZamestnanecId,\r\nTabMzdList.OdpracHod AS TabMzdList__OdpracHod,\r\nTabMzdList.HodSaz AS TabMzdList__HodSaz,\r\nTabMzdList.IdObdobi AS TabMzdList__IdObdobi\r\n FROM [dbo].TabMzdList TabMzdList) TabMzdList\r\nINNER JOIN (SELECT TabMzdObd.MzdObd_DatumOd AS TabMzdObd__MzdObd_DatumOd,\r\nTabMzdObd.MzdObd_DatumDo AS TabMzdObd__MzdObd_DatumDo,\r\nTabMzdObd.IdObdobi AS TabMzdObd__IdObdobi\r\n FROM [dbo].TabMzdObd TabMzdObd) TabMzdObd\r\nON TabMzdObd.TabMzdObd__IdObdobi = TabMzdObd.TabMzdObd__IdObdobi)"));
+
+    }
+
     // TODO:
-    //[Test]
-    //public void TestJoinProcessing()    // TODO: Test different join types and join condition operators
-    //{
-    //    ISourceEntity sourceTable1 = new SourceTable("TabMzdList", ["ZamestnanecId", "OdpracHod", "IdObdobi"]);
-    //    ISourceEntity sourceTable2 = new SourceTable("TabMzdObd", ["MzdObd_DatumOd", "MzdObd_DatumDo", "IdObdobi"]);
-    //    Join join = new(
-    //        Join.Type.Inner,
-    //        sourceTable1,
-    //        sourceTable2,
-    //        "3",
-    //        [
-    //            sourceTable1.GetColumnMapping("ZamestnanecId"),
-    //            sourceTable1.GetColumnMapping("OdpracHod"),
-    //            sourceTable2.GetColumnMapping("MzdObd_DatumOd"),
-    //            sourceTable2.GetColumnMapping("MzdObd_DatumDo"),
-    //        ],
-    //        new(JoinCondition.Operator.Equal, sourceTable1.GetColumnMapping("IdObdobi"), sourceTable2.GetColumnMapping("IdObdobi")));
 
-    //    SqlViewVisitor visitor = new("[dbo].");
-    //    join.Accept(visitor);
-    //    string actualJoinRepresentation = visitor.GetSqlView();
+    [Test]
+    public void TestEntityMappingProcessing()
+    {
+        // TODO: the view entity must generate the named view
+        SourceColumn[] sourceColumns = [
+            new("MzdObd_DatumOd", new SimpleType(SimpleType.Types.Date, false)),
+            new("MzdObd_DatumDo", new SimpleType(SimpleType.Types.Date, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ];
 
-    //    Assert.That(actualJoinRepresentation, Is.Not.Empty);
-    //    Assert.That(actualJoinRepresentation, Is.EqualTo("(SELECT TabMzdList.ZamestnanecId, TabMzdList.OdpracHod, TabMzdObd.MzdObd_DatumOd, TabMzdObd.MzdObd_DatumDo FROM (SELECT ZamestnanecId, OdpracHod, IdObdobi FROM [dbo].TabMzdList) TabMzdList INNER JOIN (SELECT MzdObd_DatumOd, MzdObd_DatumDo, IdObdobi FROM [dbo].TabMzdObd) TabMzdObd ON TabMzdList.IdObdobi = TabMzdObd.IdObdobi)"));
+        ISourceEntity sourceTable1 = new SourceTable("TabMzdList", null, [
+            new("ZamestnanecId", new SimpleType(SimpleType.Types.Integer, false)),
+            new("OdpracHod", new SimpleType(SimpleType.Types.Decimal, false)),
+            new("HodSaz", new SimpleType(SimpleType.Types.Decimal, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ]);
 
-    //}
+        ISourceEntity sourceTable2 = new SourceTable("TabMzdObd", null, [
+            new("MzdObd_DatumOd", new SimpleType(SimpleType.Types.Date, false)),
+            new("MzdObd_DatumDo", new SimpleType(SimpleType.Types.Date, false)),
+            new("IdObdobi", new SimpleType(SimpleType.Types.Integer, false)),
+        ]);
+        Join join = new(
+            Join.Type.inner,
+            sourceTable1,
+            sourceTable2,
+            "join1",
+            [
+                sourceTable1.GetColumnMapping("IdObdobi"),
+                sourceTable1.GetColumnMapping("HodSaz"),
+                sourceTable1.GetColumnMapping("OdpracHod"),
+                sourceTable1.GetColumnMapping("ZamestnanecId"),
+                sourceTable2.GetColumnMapping("IdObdobi"),
+                sourceTable2.GetColumnMapping("MzdObd_DatumDo"),
+                sourceTable2.GetColumnMapping("MzdObd_DatumOd"),
+            ],
+            new(JoinCondition.Operator.equals, sourceTable2.GetColumnMapping("IdObdobi"), sourceTable2.GetColumnMapping("IdObdobi")));
 
-    //[Test]
-    //public void TestEntityMappingProcessing()
-    //{
-    //    // TODO: the view entity must generate the named view
-    //    ISourceEntity sourceTable1 = new SourceTable("TabMzdList", ["ZamestnanecId", "OdpracHod", "IdObdobi"]);
-    //    ISourceEntity sourceTable2 = new SourceTable("TabMzdObd", ["MzdObd_DatumOd", "MzdObd_DatumDo", "IdObdobi"]);
-    //    Join join = new(
-    //        Join.Type.Inner,
-    //        sourceTable1,
-    //        sourceTable2,
-    //        "3",
-    //        [
-    //            sourceTable1.GetColumnMapping("ZamestnanecId"),
-    //            sourceTable1.GetColumnMapping("OdpracHod"),
-    //            sourceTable2.GetColumnMapping("MzdObd_DatumOd"),
-    //            sourceTable2.GetColumnMapping("MzdObd_DatumDo"),
-    //        ],
-    //        new(JoinCondition.Operator.Equal, sourceTable1.GetColumnMapping("IdObdobi"), sourceTable2.GetColumnMapping("IdObdobi")));
+        EntityMapping entityMapping = new(
+            name: "EmployeeHoursWorked2",
+            null,
+            [sourceTable1, sourceTable2, join],
+            join,
+            new() {
+                { "PersonalId", sourceTable1.GetColumnMapping("ZamestnanecId")},
+                { "HoursCount", sourceTable1.GetColumnMapping("OdpracHod")},
+                { "DateFrom", sourceTable2.GetColumnMapping("MzdObd_DatumOd")},
+                { "DateTo", null},
+                { "note", null}
+            });
+        entityMapping.SourceEntity!.AssignColumnOwnership();
 
-    //    EntityMapping entityMapping = new(
-    //        name: "EmployeeHoursWorked",
-    //        sourceEntities: [sourceTable1, sourceTable2, join],
-    //        join,
-    //        new() {
-    //            { "PersonalId", sourceTable1.GetColumnMapping("ZamestnanecId")},
-    //            { "HoursCount", sourceTable1.GetColumnMapping("OdpracHod")},
-    //            { "DateFrom", sourceTable2.GetColumnMapping("MzdObd_DatumOd")},
-    //            { "DateTo", sourceTable2.GetColumnMapping("MzdObd_DatumDo")},
-    //        });
+        SqlViewVisitor visitor = new("[dbo].");
+        entityMapping.Accept(visitor);
+        string actualJoinRepresentation = visitor.GetSqlView();
+        Console.WriteLine();
 
-    //    SqlViewVisitor visitor = new("[dbo].");
-    //    entityMapping.Accept(visitor);
-    //    string actualJoinRepresentation = visitor.GetSqlView();
-
-    //    Assert.That(actualJoinRepresentation, Is.Not.Empty);
-    //    Assert.That(actualJoinRepresentation, Is.EqualTo("CREATE VIEW [dbo].EmployeeHoursWorked AS SELECT ZamestnanecId AS PersonalId, OdpracHod AS HoursCount, MzdObd_DatumOd AS DateFrom, MzdObd_DatumDo AS DateTo FROM (SELECT TabMzdList.ZamestnanecId, TabMzdList.OdpracHod, TabMzdObd.MzdObd_DatumOd, TabMzdObd.MzdObd_DatumDo FROM (SELECT ZamestnanecId, OdpracHod, IdObdobi FROM [dbo].TabMzdList) TabMzdList INNER JOIN (SELECT MzdObd_DatumOd, MzdObd_DatumDo, IdObdobi FROM [dbo].TabMzdObd) TabMzdObd ON TabMzdList.IdObdobi = TabMzdObd.IdObdobi)"));
-    //}
+        Assert.That(actualJoinRepresentation, Is.Not.Empty);
+        Assert.That(actualJoinRepresentation, Is.EqualTo("CREATE VIEW [dbo].EmployeeHoursWorked2\r\nAS SELECT join1.TabMzdList__ZamestnanecId AS PersonalId, join1.TabMzdList__OdpracHod AS HoursCount, join1.TabMzdObd__MzdObd_DatumOd AS DateFrom, NULL AS DateTo, NULL AS note\r\nFROM (SELECT TabMzdList.TabMzdList__IdObdobi,\r\nTabMzdList.TabMzdList__HodSaz,\r\nTabMzdList.TabMzdList__OdpracHod,\r\nTabMzdList.TabMzdList__ZamestnanecId,\r\nTabMzdObd.TabMzdObd__IdObdobi,\r\nTabMzdObd.TabMzdObd__MzdObd_DatumDo,\r\nTabMzdObd.TabMzdObd__MzdObd_DatumOd\r\nFROM (SELECT TabMzdList.ZamestnanecId AS TabMzdList__ZamestnanecId,\r\nTabMzdList.OdpracHod AS TabMzdList__OdpracHod,\r\nTabMzdList.HodSaz AS TabMzdList__HodSaz,\r\nTabMzdList.IdObdobi AS TabMzdList__IdObdobi\r\n FROM [dbo].TabMzdList TabMzdList) TabMzdList\r\nINNER JOIN (SELECT TabMzdObd.MzdObd_DatumOd AS TabMzdObd__MzdObd_DatumOd,\r\nTabMzdObd.MzdObd_DatumDo AS TabMzdObd__MzdObd_DatumDo,\r\nTabMzdObd.IdObdobi AS TabMzdObd__IdObdobi\r\n FROM [dbo].TabMzdObd TabMzdObd) TabMzdObd\r\nON TabMzdObd.TabMzdObd__IdObdobi = TabMzdObd.TabMzdObd__IdObdobi) join1"));
+    }
 }
