@@ -3,4 +3,4 @@
 /// <summary>
 /// Represents the singleton service lifetime.
 /// </summary>
-public interface ISigleton;
+public interface ISingleton;
