@@ -9,13 +9,8 @@
         </div>
 
         <BIManagement.Modules.DataIntegration.Pages.DataIntegrationSubmenu />
+        <BIManagement.Modules.Deployment.Pages.DeploymentSubMenu/>
         <BIManagement.Modules.Users.Pages.UsersSubmenu/>
-
-        <div class="nav-item px-3">
-            <NavLink class="nav-link" href="weather">
-                <span class="bi bi-list-nested-nav-menu" aria-hidden="true"></span> Weather
-            </NavLink>
-        </div>
 
         <AuthorizeView>
             <Authorized>
@@ -35,15 +30,6 @@
                 </div>
             </Authorized>
             <NotAuthorized>
-                <div class="nav-item px-3">
-                    <NavLink class="nav-link" href="Account/
-
-
-
-">
-                        <span class="bi bi-person-nav-menu" aria-hidden="true"></span> Register
-                    </NavLink>
-                </div>
                 <div class="nav-item px-3">
                     <NavLink class="nav-link" href="Account/Login">
                         <span class="bi bi-person-badge-nav-menu" aria-hidden="true"></span> Login
