@@ -7,4 +7,4 @@
 @using static Microsoft.AspNetCore.Components.Web.RenderMode
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.JSInterop
-@using ManagementApp.Components
+@using BIManagement.ManagementApp.Components
