@@ -8,25 +8,32 @@ namespace BIManagement.Modules.DataIntegration.Domain.Mapping;
 public interface ISchemaMappingRepository
 {
     /// <summary>
-    /// Gets all schema mappings of a given costumer.
+    /// Gets all schema mappings of a given customer.
     /// </summary>
-    /// <param name="costumerId">Id of the respective costumer</param>
+    /// <param name="customerId">Id of the respective customer</param>
     /// <returns>
     /// Task object that represents the asynchronous operation.
     /// The Value property returns instance of <see cref="IEnumerable{SchemaMapping}"/> on success.
     /// </returns>
-    Task<IReadOnlyList<SchemaMapping>> GetSchemaMappings(string costumerId);
+    Task<IReadOnlyList<SchemaMapping>> GetSchemaMappings(string customerId);
 
     /// <summary>
-    /// Gets a schema mapping of a given costumer and target database table.
+    /// Deletes the schema mapping of a given customer.
     /// </summary>
-    /// <param name="costumerId">Id of the respective costumer </param>
+    /// <param name="customerId">Id of the respective customer</param>
+    /// <returns>Task object that represents the asynchronous operation.</returns>
+    Task DeleteSchemaMappings(string customerId);
+
+    /// <summary>
+    /// Gets a schema mapping of a given customer and target database table.
+    /// </summary>
+    /// <param name="customerId">Id of the respective customer</param>
     /// <param name="targetDbTableId">Id of the respective table from the target database.</param>
     /// <returns>
     /// Task object that represents the asynchronous operation.
     /// The Value property returns instance of <see cref="SchemaMapping"/> on success otherwise a null.
     /// </returns>
-    Task<SchemaMapping?> GetSchemaMapping(string costumerId, int targetDbTableId);
+    Task<SchemaMapping?> GetSchemaMapping(string customerId, int targetDbTableId);
 
     /// <summary>
     /// Asynchronously adds or updates the given instance of <see cref="SchemaMapping"/>.
