@@ -17,14 +17,14 @@ namespace BIManagement.Modules.DataIntegration.Persistence.Migrations
 #pragma warning disable 612, 618
             modelBuilder
                 .HasDefaultSchema("dataIntegration")
-                .HasAnnotation("ProductVersion", "8.0.6")
+                .HasAnnotation("ProductVersion", "9.0.0-preview.5.24306.3")
                 .HasAnnotation("Relational:MaxIdentifierLength", 128);
 
             SqlServerModelBuilderExtensions.UseIdentityColumns(modelBuilder);
 
-            modelBuilder.Entity("BIManagement.Modules.DataIntegration.Domain.DatabaseConnection.CostumerDbConnectionConfiguration", b =>
+            modelBuilder.Entity("BIManagement.Modules.DataIntegration.Domain.DatabaseConnection.CustomerDbConnectionConfiguration", b =>
                 {
-                    b.Property<string>("CostumerId")
+                    b.Property<string>("CustomerId")
                         .HasMaxLength(450)
                         .HasColumnType("nvarchar(450)");
 
@@ -35,14 +35,14 @@ namespace BIManagement.Modules.DataIntegration.Persistence.Migrations
                     b.Property<int>("Provider")
                         .HasColumnType("int");
 
-                    b.HasKey("CostumerId");
+                    b.HasKey("CustomerId");
 
                     b.ToTable("DatabaseConnectionConfigurations", "dataIntegration");
                 });
 
-            modelBuilder.Entity("BIManagement.Modules.DataIntegration.Domain.DbModelling.CostumerDbModel", b =>
+            modelBuilder.Entity("BIManagement.Modules.DataIntegration.Domain.DbModelling.CustomerDbModel", b =>
                 {
-                    b.Property<string>("CostumerId")
+                    b.Property<string>("CustomerId")
                         .HasMaxLength(450)
                         .HasColumnType("nvarchar(450)");
 
@@ -50,14 +50,14 @@ namespace BIManagement.Modules.DataIntegration.Persistence.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.HasKey("CostumerId");
+                    b.HasKey("CustomerId");
 
-                    b.ToTable("CostumerDbModels", "dataIntegration");
+                    b.ToTable("CustomerDbModels", "dataIntegration");
                 });
 
             modelBuilder.Entity("BIManagement.Modules.DataIntegration.Domain.Mapping.SchemaMapping", b =>
                 {
-                    b.Property<string>("CostumerId")
+                    b.Property<string>("CustomerId")
                         .HasMaxLength(450)
                         .HasColumnType("nvarchar(450)");
 
@@ -71,7 +71,7 @@ namespace BIManagement.Modules.DataIntegration.Persistence.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.HasKey("CostumerId", "TargetDbTableId");
+                    b.HasKey("CustomerId", "TargetDbTableId");
 
                     b.ToTable("SchemaMappings", "dataIntegration");
                 });
@@ -105,21 +105,21 @@ namespace BIManagement.Modules.DataIntegration.Persistence.Migrations
                         {
                             Id = 1,
                             Schema = "dbo",
-                            TableModel = "{\"name\":\"Employees\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"Id\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"Id\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}},{\"name\":\"PersonalID\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}},{\"name\":\"ExternalId\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}},{\"name\":\"FirstName\",\"dataType\":{\"$type\":\"nVarCharMax\",\"isNullable\":false}},{\"name\":\"Lastname\",\"dataType\":{\"$type\":\"nVarCharMax\",\"isNullable\":false}}],\"description\":null}",
+                            TableModel = "{\"name\":\"Employees\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"Id\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"Id\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}},{\"name\":\"PersonalID\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}},{\"name\":\"ExternalId\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}},{\"name\":\"FirstName\",\"dataType\":{\"type\":\"nVarCharMax\",\"isNullable\":false}},{\"name\":\"Lastname\",\"dataType\":{\"type\":\"nVarCharMax\",\"isNullable\":false}}],\"description\":null}",
                             TableName = "Employees"
                         },
                         new
                         {
                             Id = 2,
                             Schema = "dbo",
-                            TableModel = "{\"name\":\"Workplaces\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"Id\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"Id\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}},{\"name\":\"Label\",\"dataType\":{\"$type\":\"nVarCharMax\",\"isNullable\":false}},{\"name\":\"Name\",\"dataType\":{\"$type\":\"nVarCharMax\",\"isNullable\":true}}],\"description\":null}",
+                            TableModel = "{\"name\":\"Workplaces\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"Id\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"Id\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}},{\"name\":\"Label\",\"dataType\":{\"type\":\"nVarCharMax\",\"isNullable\":false}},{\"name\":\"Name\",\"dataType\":{\"type\":\"nVarCharMax\",\"isNullable\":true}}],\"description\":null}",
                             TableName = "Workplaces"
                         },
                         new
                         {
                             Id = 3,
                             Schema = "dbo",
-                            TableModel = "{\"name\":\"WorkReports\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"ID\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"ID\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":false}},{\"name\":\"OrderId\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":true}},{\"name\":\"ProductionOperationId\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":true}},{\"name\":\"Quantity\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Decimal\",\"isNullable\":false}},{\"name\":\"ExpectedTime\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Decimal\",\"isNullable\":false}},{\"name\":\"DateTime\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Datetime\",\"isNullable\":false}},{\"name\":\"ProductType\",\"dataType\":{\"$type\":\"nVarCharMax\",\"isNullable\":true}},{\"name\":\"WorkerId\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":true}},{\"name\":\"WorkPlaceId\",\"dataType\":{\"$type\":\"simple\",\"type\":\"Integer\",\"isNullable\":true}}],\"description\":null}",
+                            TableModel = "{\"name\":\"WorkReports\",\"schema\":\"dbo\",\"primaryKeys\":[{\"name\":\"ID\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}}],\"columns\":[{\"name\":\"ID\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":false}},{\"name\":\"OrderId\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":true}},{\"name\":\"ProductionOperationId\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":true}},{\"name\":\"Quantity\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Decimal\",\"isNullable\":false}},{\"name\":\"ExpectedTime\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Decimal\",\"isNullable\":false}},{\"name\":\"DateTime\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Datetime\",\"isNullable\":false}},{\"name\":\"ProductType\",\"dataType\":{\"type\":\"nVarCharMax\",\"isNullable\":true}},{\"name\":\"WorkerId\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":true}},{\"name\":\"WorkPlaceId\",\"dataType\":{\"type\":\"simple\",\"simpleType\":\"Integer\",\"isNullable\":true}}],\"description\":null}",
                             TableName = "WorkReports"
                         });
                 });
