@@ -1,13 +1,14 @@
-﻿namespace BIManagement.Modules.Deployment.Persistence.Constants;
+﻿using BIManagement.Modules.Deployment.Domain;
+
+namespace BIManagement.Modules.Deployment.Persistence.Constants;
 
 /// <summary>
 /// Encapsulates the table names used in the Deployment module.
 /// </summary>
 internal static class TableNames
 {
-    ///// <summary>
-    ///// </summary>
-    //public const string Deployment = "Deployments";
-
-    // TODO: Add more table names here
+    /// <summary>
+    /// Name of the table for <see cref="MetabaseDeployment"/>
+    /// </summary>
+    public const string MetabaseDeployments = "MetabaseDeployments";
 }
