@@ -7,14 +7,14 @@ using Microsoft.AspNetCore.Authorization;
 namespace BIManagement.Modules.Users.Api.Authorization.PolicyAttributes;
 
 /// <summary>
-/// Custom authorization attribute that allows only admins to access the method.
+/// Custom authorization attribute that allows only customers to access the method.
 /// </summary>
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = true, Inherited = true)]
-public class AuthorizeOnlyCostumersAttribute : AuthorizeAttribute
+public class AuthorizeOnlyCustomersAttribute : AuthorizeAttribute
 {
-    public AuthorizeOnlyCostumersAttribute()
+    public AuthorizeOnlyCustomersAttribute()
     {
-        Roles = Domain.Roles.Costumer;
+        Roles = Domain.Roles.Customer;
     }
 }
 
