@@ -47,8 +47,7 @@ namespace BIManagement.Modules.Users.Infrastructure.ServiceInstallers
                 .AddSignInManager()
                 .AddDefaultTokenProviders();
 
-            // TODO: consider using only the notification email sender
-            services.AddSingleton<IEmailSender<ApplicationUser>, Identity.IdentityNoOpEmailSender>();
+            services.AddSingleton<IEmailSender<ApplicationUser>, Identity.IdentityEmailSender>();
             services.AddHostedService<SeedingStartupTask>();
         }
     }
