@@ -53,7 +53,7 @@
             NavigationManager.ToAbsoluteUri("Account/ResetPassword").AbsoluteUri,
             new Dictionary<string, object?> { ["code"] = code });
 
-        await EmailSender.SendPasswordResetLinkAsync(user, Input.Email, HtmlEncoder.Default.Encode(callbackUrl));
+        await EmailSender.SendPasswordResetLinkAsync(user, Input.Email, callbackUrl);
 
         RedirectManager.RedirectTo("Account/ForgotPasswordConfirmation");
     }
