@@ -89,7 +89,7 @@
             new Dictionary<string, object?> { ["userId"] = userId, ["email"] = Input.NewEmail, ["code"] = code });
 
         // TODO: MOVE THIS TO USER MANAGER
-        await EmailSender.SendConfirmationLinkAsync(user, Input.NewEmail, HtmlEncoder.Default.Encode(callbackUrl));
+        await EmailSender.SendConfirmationLinkAsync(user, Input.NewEmail, callbackUrl);
 
         message = "Confirmation link to change email sent. Please check your email.";
     }
@@ -108,7 +108,7 @@
             NavigationManager.ToAbsoluteUri("Account/ConfirmEmailAndSetPassword").AbsoluteUri,
             new Dictionary<string, object?> { ["userId"] = userId, ["code"] = code });
 
-        await EmailSender.SendConfirmationLinkAsync(user, email, HtmlEncoder.Default.Encode(callbackUrl));
+        await EmailSender.SendConfirmationLinkAsync(user, email, callbackUrl);
 
         message = "Verification email sent. Please check your email.";
     }
