@@ -48,7 +48,7 @@
 }
 else
 {
-    <p>Costumer not found.</p>
+    <p>Customer not found.</p>
 }
 
 @code {
